+++ WireBrainCoffee.ShopInfoTool/WireBrainCoffee.ShopInfoTool/IHelpCommandHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace WireBrainCoffee.ShopInfoTool
+{
+    internal interface IHelpCommandHandler
+    {
+        void HandleCommand();
+    }
+}
