+++ WireBrainCoffee.ShopInfoTool/WireBrainCoffee.ShopInfoTool/Program.cs
@@ -18,44 +18,28 @@ namespace WireBrainCoffee.ShopInfoTool
                 var line = Console.ReadLine();
                 var coffeeShops = coffeeShopDataProvider.LoadCoffeeShops();
 
+
                 if (string.Equals("quit", line, StringComparison.OrdinalIgnoreCase))
                 {
                     break;
                 }
+                var commandHandler = string.Equals("help", line, StringComparison.OrdinalIgnoreCase)
+                    ? new HelpCommandHandler(coffeeShops) as IHelpCommandHandler
+                    : new CoffeeShopCommandHandler(coffeeShops, line);
+
+                commandHandler.HandleCommand();
 
 
-                if (string.Equals("help", line, StringComparison.OrdinalIgnoreCase))
-                {
-                    Console.WriteLine("> Available coffee shop commands:");
-                    foreach (var coffeeShop in coffeeShops)
-                    {
-                        Console.WriteLine($">" + coffeeShop.Location);
-                    }
-                }
-                else
-                {
-                    var foundCoffeeShop = coffeeShops
-                        .Where(x => x.Location.StartsWith(line, StringComparison.OrdinalIgnoreCase))
-                        .ToList();
-                    if (foundCoffeeShop.Count == 0)
-                    {
-                        Console.WriteLine($"> Command '{line}' not found ");
-                    }
-                    else if (foundCoffeeShop.Count == 1)
-                    {
-                        var coffeeShop = foundCoffeeShop.Single();
-                        Console.WriteLine($"> Location: {coffeeShop.Location}");
-                        Console.WriteLine($"> Beans in stock: {coffeeShop.BeansInStockInKg}");
-                    }
-                    else
-                    {
-                        Console.WriteLine($"> Multiple matching coffee shop commands found:");
-                        foreach (var coffeeType in foundCoffeeShop)
-                        {
-                            Console.WriteLine($"> {coffeeType.Location}");
-                        }
-                    }
-                }
+                //if (string.Equals("help", line, StringComparison.OrdinalIgnoreCase))
+                //{
+                //    commandHandler = new HelpCommandHandler(coffeeShops);
+                //    commandHandler.HandleCommand();
+                //}
+                //else
+                //{
+                //    commandHandler = new CoffeeShopCommandHandler(coffeeShops, line);
+                //    commandHandler.HandleCommand();
+                //}
             }
         }
     }
